+++ src/main/java/io/github/egor78945/weatherSDK/enumeration/SdkMode.java
@@ -1,0 +1,15 @@
+package io.github.egor78945.weatherSDK.enumeration;
+
+/**
+ * Represents SDK work mode types.
+ */
+public enum SdkMode {
+    /**
+     * Cached weather data updates by user request.
+     */
+    ON_DEMAND,
+    /**
+     * Cached weather data updates automatically by individual thread.
+     */
+    POLLING
+}
