+++ src/main/java/io/github/egor78945/weatherSDK/model/dto/CacheNode.java
@@ -1,0 +1,35 @@
+package io.github.egor78945.weatherSDK.model.dto;
+
+import java.time.LocalDateTime;
+
+/**
+ * Class for structure cache information.
+ * @param <T> A class type for {{@link #payload}}.
+ */
+public abstract class CacheNode <T> {
+    /**
+     * Time of creating {@link CacheNode}.
+     */
+    protected final LocalDateTime localDateTime;
+    /**
+     * Caching object.
+     */
+    protected final T payload;
+
+    /**
+     * Creating default {@link CacheNode}.
+     * @param payload Cashing object.
+     */
+    public CacheNode(T payload) {
+        this.localDateTime = LocalDateTime.now();
+        this.payload = payload;
+    }
+
+    public LocalDateTime getLocalDateTime() {
+        return localDateTime;
+    }
+
+    public T getPayload() {
+        return payload;
+    }
+}
