--- src/main/java/org/example/weather/enumeration/SdkMode.java
@@ -1,15 +1,0 @@
-package org.example.weather.enumeration;
-
-/**
- * Represents SDK work mode types.
- */
-public enum SdkMode {
-    /**
-     * Cached weather data updates by user request.
-     */
-    ON_DEMAND,
-    /**
-     * Cached weather data updates automatically by individual thread.
-     */
-    POLLING
-}
