--- src/main/java/org/example/weather/mapper/JsonMapper.java
@@ -1,9 +1,0 @@
-package org.example.weather.mapper;
-
-/**
- * Mapper for JSON HTTP response.
- */
-public interface JsonMapper <T> {
-    T mapTo(String json);
-    String mapTo(T object);
-}
