--- src/main/resources/application.properties
@@ -1,5 +1,0 @@
-weather.secret=375503d1cb8283ef534d1941f9bd901d
-weather.mode=on_demand
-weather.cache.expiration=1
-weather.cache.update=1
-weather.cache.limit=10
